+++ doc/release-notes.md
@@ -7,3 +7,5 @@ This release includes the following features and fixes:
    the bitcoin-qt GUI.
  - The default wallet will now be labeled `[default wallet]` in the bitcoin-qt
    GUI if no name is provided by the `-wallet` option on start up.
+ - It is now possible to unload wallets dynamically at runtime. This feature is
+   currently only available through the RPC interface.